memberships added

+++ manager/src/Model/Work/UseCase/Projects/Role/Remove/Handler.php
@@ -5,6 +5,7 @@ namespace App\Model\Work\UseCase\Projects\Role\Remove;
 
 
 use App\Model\Flusher;
+use App\Model\Work\Entity\Projects\Project\ProjectRepository;
 use App\Model\Work\Entity\Projects\Role\Id;
 use App\Model\Work\Entity\Projects\Role\RoleRepository;
 
@@ -22,16 +23,22 @@ class Handler
      * @var Flusher
      */
     private $flusher;
+    /**
+     * @var ProjectRepository
+     */
+    private $projectRepository;
 
     /**
      * Handler constructor.
      * @param RoleRepository $roleRepository
+     * @param ProjectRepository $projectRepository
      * @param Flusher $flusher
      */
-    public function __construct(RoleRepository $roleRepository, Flusher $flusher)
+    public function __construct(RoleRepository $roleRepository, ProjectRepository $projectRepository, Flusher $flusher)
     {
         $this->roleRepository = $roleRepository;
         $this->flusher = $flusher;
+        $this->projectRepository = $projectRepository;
     }
 
     /**
@@ -40,6 +47,10 @@ class Handler
     public function handle(Command $command)
     {
         $role = $this->roleRepository->get(new Id($command->id));
+
+        if ($this->projectRepository->hasMemberWithRole($role->getId())) {
+            throw new \DomainException('Role contains member.');
+        }
 
         $this->roleRepository->remove($role);
